refactored file names, removed dead code
