+++ gcal/calendar.go
@@ -5,37 +5,30 @@ package gcal
 
 import (
 	"context"
-	"net/http"
 
 	"github.com/pkg/errors"
 	"google.golang.org/api/calendar/v3"
 	"google.golang.org/api/option"
 
 	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/remote"
-	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/utils/bot"
 )
 
-func (c *client) GetCalendars(remoteUserID string) ([]*remote.Calendar, error) {
-	if true {
-		return nil, errors.New("gcal GetCalendars not implemented")
-	}
-
-	var v struct {
-		Value []*remote.Calendar `json:"value"`
-	}
-	req := c.rbuilder.Users().ID(remoteUserID).Calendars().Request()
-	req.Expand("children")
-	err := req.JSONRequest(c.ctx, http.MethodGet, "", nil, &v)
-	if err != nil {
-		return nil, errors.Wrap(err, "msgraph GetCalendars")
-	}
-	c.Logger.With(bot.LogContext{
-		"UserID": remoteUserID,
-		"v":      v.Value,
-	}).Infof("gcal: GetUserCalendars returned `%d` calendars.", len(v.Value))
-	return v.Value, nil
+// CreateCalendar creates a calendar
+func (c *client) CreateCalendar(remoteUserID string, calIn *remote.Calendar) (*remote.Calendar, error) {
+	return nil, errors.New("gcal CreateCalendar not implemented")
 }
 
+// DeleteCalendar deletes a calendar
+func (c *client) DeleteCalendar(remoteUserID string, calID string) error {
+	return errors.New("gcal DeleteCalendar not implemented")
+}
+
+// GetCalendars returns a list of calendars
+func (c *client) GetCalendars(remoteUserID string) ([]*remote.Calendar, error) {
+	return nil, errors.New("gcal GetCalendars not implemented")
+}
+
+// GetDefaultCalendar returns the default calendar for the user
 func (c *client) GetDefaultCalendar() (*remote.Calendar, error) {
 	service, err := calendar.NewService(context.Background(), option.WithHTTPClient(c.httpClient))
 	if err != nil {
@@ -53,6 +46,7 @@ func (c *client) GetDefaultCalendar() (*remote.Calendar, error) {
 	return remoteCal, nil
 }
 
+// convertGoogleCalendarToRemoteCalendar converts a google calendar to a local representation calendar
 func convertGoogleCalendarToRemoteCalendar(cal *calendar.Calendar) *remote.Calendar {
 	return &remote.Calendar{
 		ID:           cal.Id,
