--- gcal/get_super_user_token.go
@@ -1,12 +1,0 @@
-// Copyright (c) 2019-present Mattermost, Inc. All Rights Reserved.
-// See License for license information.
-
-package gcal
-
-import (
-	"github.com/mattermost/mattermost-plugin-mscalendar/calendar/remote"
-)
-
-func (c *client) GetSuperuserToken() (string, error) {
-	return "", remote.ErrNotImplemented
-}
